refactor: move placement related functions out.
