+++ packages/ui/src/composables/index.ts
@@ -63,3 +63,4 @@ export * from './useSlotPassed'
 export * from './useReactiveComputed'
 export * from './useIcon'
 export * from './useGlobalConfig'
+export * from './usePlacement'
