+++ packages/ui/src/composables/useDropdown.ts
@@ -3,47 +3,14 @@ import { computed, Ref, unref, watchPostEffect } from 'vue'
 import { mapObject } from '../utils/map-object'
 import { useDomRect } from './useDomRect'
 import { useDocument } from './useDocument'
+import { Placement, PlacementAlignment, parsePlacementWithAlias } from './usePlacement'
 
-export type PlacementPosition = 'top' | 'bottom' | 'left' | 'right'
-export type PlacementAlignment = 'start' | 'end' | 'center'
-export type Placement = PlacementPosition | 'auto' | `${PlacementPosition}-${PlacementAlignment}`
-export type PlacementAlias = 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right'
-export type PlacementWithAlias = Placement | PlacementAlias
 export type Offset = number | [number, number]
 
 type Coords = { x: number, y: number }
 type AlignCoords = { main: number, cross: number }
 
-export const placementsPositions = ['top', 'bottom', 'left', 'right']
-  .reduce((acc, position) => [...acc, position, `${position}-start`, `${position}-end`, `${position}-center`], ['auto'] as string[])
-
-export const placementAliasesPositions = ['top', 'bottom']
-  .flatMap(position => [`${position}-left`, `${position}-right`])
-
-export const placementsPositionsWithAliases = [...placementsPositions, ...placementAliasesPositions]
-
 const coordsToCss = ({ x, y }: Coords) => ({ left: `${x}px`, top: `${y}px` })
-
-export const aliasToPlacement: Record<string, Placement> = {
-  'top-left': 'top-start',
-  'top-right': 'top-end',
-  'bottom-left': 'bottom-start',
-  'bottom-right': 'bottom-end',
-}
-
-const parsePlacementWithAlias = (placementWithAlias: PlacementWithAlias) => {
-  const placement = aliasToPlacement[placementWithAlias] || placementWithAlias
-  return parsePlacement(placement)
-}
-
-const parsePlacement = (placement: Placement) => {
-  let [position, align] = placement.split('-') as [PlacementPosition | 'auto', PlacementAlignment | undefined]
-  if (position === 'auto') {
-    position = 'bottom'
-  }
-
-  return { position, align: align || 'center' }
-}
 
 const parseOffset = (offset: Offset): AlignCoords => {
   return Array.isArray(offset) ? { main: offset[0], cross: offset[1] } : { main: offset, cross: 0 }
